+++ src/app/Data/services/forecast-data/forecast-data.service.ts
@@ -1,0 +1,16 @@
+import { Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
+import { Observable } from 'rxjs';
+
+import { environment } from '../../../../environments/environment';
+import { ForecastData } from '../../../Domain/models/forecast-data.model';
+
+@Injectable({ providedIn: 'root' })
+export class ForecastDataService {
+  constructor(private http: HttpClient) {
+  }
+
+  getForestData(id: number): Observable<ForecastData> {
+    return this.http.get<ForecastData>(`${environment.dataPath}data_${id}.json`);
+  }
+}