+++ src/app/Domain/state/forecast-data/forecast-data.effect.ts
@@ -1,11 +1,9 @@
 import { Actions, createEffect, ofType } from '@ngrx/effects';
 import { exhaustMap, map } from 'rxjs';
 import { Injectable } from '@angular/core';
-import { Store } from '@ngrx/store';
 
 import * as ForecastActions from './forecast-data.actions';
-import { AppState } from '../app.state';
-import { ForecastDataService } from '../../../Data/services/forecast-data.service';
+import { ForecastDataService } from '../../../Data/services/forecast-data/forecast-data.service';
 import { ForecastData } from '../../models/forecast-data.model';
 
 @Injectable({ providedIn: 'root' })
@@ -18,14 +16,14 @@ export class ForecastDataEffect {
           return this.forecastService
             .getForestData(forecastId)
             .pipe(
-              map((data: ForecastData) => {
-                return ForecastActions.getForecastData({ forecastId, forecastData: data })
+              map((forecastData: ForecastData) => {
+                return ForecastActions.getForecastData({ forecastId, forecastData });
               })
             )
         })
       )
   );
 
-  constructor(private actions$: Actions, private store: Store<AppState>, private forecastService: ForecastDataService) {
+  constructor(private actions$: Actions, private forecastService: ForecastDataService) {
   }
 }